requirements.txt
nemo_toolkit[asr]
sounddevice
numpy
scipy
torch
torchaudio
webrtcvad
psycopg2-binary
python-dotenv
